modified line 61 in sibdrv_interp.F90 from
(sib(i)%stat%coszbar.ne.0.0_dbl_kind)

to 

(sib(i)%stat%coszbar > 0.01_dbl_kind)

There was a particular pathology that arose when coszbar was very small 
and cosz small that resulted in radiation being scaled by a very large 
number. 

This came up in global (6-hour data period) simulations: for a few 
points, coszbar would be very small because the entire 6-hour 
period-except for a couple of timesteps-had a cosz < 0. Then, when the 
sun peeked above the horizon for those last couple of timeteps in the 
data period, the radiation value was scaled by a huge number and model 
wackiness ensued. We're hoping 0.01 will fix it, but we need to do more 
research on it.