New branch to hold the latest version of the code: This code has the following updates
1) Reto's energy/moisture balance checks
2) Andrew Philpott's mapper changes; interpolation of NDVI and ability to 
utilize flexible NDVI timing
3) Autotrophic respiration
4) New feature: Ability to run simulations for a tower (local NDVI) with large 
scale (NCEP) meteorology

The idea is that this code will replace the main trunk. However, there is one 
aspect of this code that I am uncomfortable with: The flexibility of the 
single-point ASCII parameter file has been removed. There is now a TI/Tvarying 
set of parameter files for a single-point run, like for other runs. This is 
consistent, but a lot of the flexibility I built into the multi-phys has been 
removed. I won't make this branch the main trunk until I resolve this problem.