Ian found a problem with co2s:
The reason for this is that molecules of different sizes can behave differently. During purely turbulent processes (in the canopy air space for example, where model fluxes are influenced by ra) there is no difference between water and CO2, and they will have a ratio of 1.0. For molecular diffusion through the stomate, the size ratio is 1.6, and for mixed turbulent-molecular transport (through rb) the ratio is an intermediate value of 1.4.

If you look at the code you can see that the 1.6 term is used when we determine pco2i in the sortin/cycalc loop (~line 508), but had not been used when we back out pco2s after determining the final value for pco2i.